pyproject.toml
[tool.poetry]
name = "hw-django-30-1"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "hw_django_30"}]

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.2.4"
djangorestframework = "^3.16.0"
pillow = "^11.3.0"
psycopg2 = "^2.9.10"
python-dotenv = "^1.1.1"
django-filter = "^25.1"
djangorestframework-simplejwt = "^5.5.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
